move assistant to separate file
